default profile pic support and publication process added

+++ main_api/models.py
@@ -1,3 +1,4 @@
+from django.core.files import File
 from django.db import models
 import os
 from django.dispatch import receiver
@@ -27,7 +28,7 @@ class People(models.Model):
     description = models.CharField(max_length=200)
     web_page = models.CharField(max_length=500)
     join_date = models.DateTimeField('Join Date')
-    image = models.ImageField(upload_to='people_image', blank=True)
+    image = models.ImageField(upload_to='people_image', blank=True, default='people_image/default-pic.jpg')
 
     def __str__(self):
         return self.first_name + ' ' + self.last_name
@@ -39,7 +40,7 @@ def auto_delete_file_on_delete(sender, instance, **kwargs):
     Deletes file from filesystem
     when corresponding `MediaFile` object is deleted.
     """
-    if instance.image:
+    if instance.image and instance.image.name != 'people_image/default-pic.jpg':
         if os.path.isfile(instance.image.path):
             os.remove(instance.image.path)
 
@@ -58,8 +59,11 @@ def auto_delete_file_on_change(sender, instance, **kwargs):
         old_file = People.objects.get(pk=instance.pk).image
     except People.DoesNotExist:
         return False
+    if instance.image.name == '':
+        instance.image = 'people_image/default-pic.jpg'
 
     new_file = instance.image
-    if not old_file == new_file:
-        if os.path.isfile(old_file.path):
-            os.remove(old_file.path)
+    if old_file.name != "" and old_file.name != 'people_image/default-pic.jpg':
+        if not old_file == new_file:
+            if os.path.exists(old_file.path):
+                os.remove(old_file.path)
